Build: Move the build tool to a dedicated package
