+++ package.json
@@ -33,7 +33,6 @@
 		"@babel/plugin-syntax-jsx": "7.25.7",
 		"@babel/runtime-corejs3": "7.25.7",
 		"@babel/traverse": "7.25.7",
-		"@emotion/babel-plugin": "11.11.0",
 		"@emotion/is-prop-valid": "1.2.2",
 		"@emotion/jest": "11.7.1",
 		"@emotion/native": "11.0.0",
@@ -83,9 +82,9 @@
 		"@types/semver": "7.3.8",
 		"@types/sprintf-js": "1.1.2",
 		"@types/uuid": "8.3.1",
+		"@wordpress/build": "file:./packages/wp-build",
 		"ajv": "8.7.1",
 		"appium": "2.0.0",
-		"autoprefixer": "10.4.21",
 		"babel-jest": "29.7.0",
 		"babel-loader": "9.2.1",
 		"babel-plugin-inline-json-import": "0.3.2",
@@ -94,11 +93,9 @@
 		"babel-plugin-transform-remove-console": "6.9.4",
 		"benchmark": "2.1.4",
 		"browserslist": "4.25.1",
-		"browserslist-to-esbuild": "2.1.1",
 		"caniuse-lite": "1.0.30001731",
 		"chalk": "4.1.1",
 		"change-case": "4.1.2",
-		"chokidar": "^4.0.0",
 		"client-zip": "^2.4.5",
 		"commander": "9.2.0",
 		"concurrently": "3.5.0",
@@ -106,12 +103,8 @@
 		"core-js-builder": "3.39.0",
 		"cross-env": "7.0.3",
 		"css-loader": "6.2.0",
-		"cssnano": "6.0.1",
 		"deep-freeze": "0.0.1",
 		"equivalent-key-map": "0.2.2",
-		"esbuild": "0.25.10",
-		"esbuild-plugin-babel": "0.2.3",
-		"esbuild-sass-plugin": "3.3.1",
 		"escape-html": "1.0.3",
 		"eslint-import-resolver-node": "0.3.4",
 		"eslint-plugin-eslint-comments": "3.1.2",
@@ -124,7 +117,7 @@
 		"eslint-plugin-storybook": "0.9.0",
 		"eslint-plugin-testing-library": "6.0.2",
 		"execa": "4.0.2",
-		"fast-glob": "3.2.7",
+		"fast-glob": "^3.2.7",
 		"filenamify": "4.2.0",
 		"glob": "7.1.2",
 		"husky": "7.0.0",
@@ -144,10 +137,8 @@
 		"node-fetch": "2.7.0",
 		"npm-run-all": "4.1.5",
 		"patch-package": "8.0.0",
-		"postcss": "8.4.38",
 		"postcss-loader": "6.2.1",
 		"postcss-local-keyframes": "^0.0.2",
-		"postcss-modules": "6.0.1",
 		"prettier": "npm:wp-prettier@3.0.3",
 		"progress": "2.0.3",
 		"puppeteer-core": "23.10.1",
@@ -163,7 +154,6 @@
 		"redux": "5.0.1",
 		"resize-observer-polyfill": "1.5.1",
 		"rimraf": "5.0.10",
-		"rtlcss": "4.3.0",
 		"sass": "1.54.0",
 		"sass-embedded": "1.93.2",
 		"sass-loader": "16.0.3",
@@ -178,7 +168,6 @@
 		"style-loader": "3.2.1",
 		"terser": "5.32.0",
 		"terser-webpack-plugin": "5.3.10",
-		"toposort": "2.0.2",
 		"typescript": "5.7.2",
 		"uuid": "9.0.1",
 		"wait-on": "8.0.1",
