+++ packages/wp-build/CHANGELOG.md
@@ -1,0 +1,12 @@
+<!-- Learn how to maintain this file at https://github.com/WordPress/gutenberg/tree/HEAD/packages#maintaining-changelogs. -->
+
+## Unreleased
+
+### New Features
+
+- Initial release of `@wordpress/build` package
+- Transpilation support for TypeScript/JSX to CommonJS and ESM formats
+- SCSS and CSS modules compilation with LTR and RTL support
+- WordPress script and module bundling
+- Automatic PHP registration file generation
+- Watch mode for development
