+++ packages/wp-build/src/build.mjs
@@ -1,5 +1,8 @@
 #!/usr/bin/env node
 
+// console should be allowed in a build tool.
+/* eslint-disable no-console */
+
 /**
  * External dependencies
  */
@@ -28,13 +31,13 @@ import { groupByDepth, findScriptsToRebundle } from './dependency-graph.mjs';
 import {
 	generatePhpFromTemplate,
 	getPhpReplacements,
-} from './lib/php-generator.mjs';
+} from './php-generator.mjs';
 import {
 	getPackageInfo,
 	getPackageInfoFromFile,
 	kebabToCamelCase,
-} from './lib/package-utils.mjs';
-import { createWordpressExternalsPlugin } from './lib/wordpress-externals-plugin.mjs';
+} from './package-utils.mjs';
+import { createWordpressExternalsPlugin } from './wordpress-externals-plugin.mjs';
 
 const ROOT_DIR = process.cwd();
 const PACKAGES_DIR = path.join( ROOT_DIR, 'packages' );
@@ -1328,3 +1331,4 @@ main().catch( ( error ) => {
 	console.error( '❌ Build failed:', error );
 	process.exit( 1 );
 } );
+/* eslint-enable no-console */
