+++ packages/wp-build/src/php-generator.mjs
