+++ packages/wp-build/src/wordpress-externals-plugin.mjs
