+++ packages/wp-build/templates/index.php.template
