+++ packages/wp-build/templates/module-registration.php.template
