+++ packages/wp-build/templates/module-registry.php.template
