+++ packages/wp-build/templates/script-registration.php.template
