+++ packages/wp-build/templates/script-registry.php.template
