+++ packages/wp-build/templates/style-registration.php.template
