+++ packages/wp-build/templates/style-registry.php.template
