+++ packages/wp-build/templates/version.php.template
