+++ tsconfig.json
@@ -66,6 +66,7 @@
 		{ "path": "packages/vips" },
 		{ "path": "packages/warning" },
 		{ "path": "packages/wordcount" },
+		{ "path": "packages/wp-build" },
 		{ "path": "test/e2e" },
 		{ "path": "packages/jest-console" },
 		{ "path": "test/performance" }
